// resources/rpt-app/src/app/interfaces/landTaxTable.ts
export interface landTaxTable {
   id: string;
   arpNo: string;
   pin: string;
   surveyNo: string;
   lotNo: string;
   blockNo: string;
   streetNo: string;
   brgy: string;
   subd: string;
   city: string;
   province: string;
   class: string;
   subclass: string;
   area: string;
   assessedVal: string;
   stat: string;
}