// resources/rpt-app/src/app/interfaces/landTaxTableBldg.ts
export interface landTaxTableBldg {
   id: string;
   arpNo: string;
   pin: string;
   brgy: string;
   subd: string;
   city: string;
   province: string;
   kind: string;
   structType: string;
   bldgPermit: string;
   dateConstr: string;
   storey: string;
   actualUse: string;
   assessedVal: string;
}