// resources/rpt-app/src/app/classes/lTaxClearance.ts
export class lTaxClearance {
   current_date: string;
   owner_names: string;
   pin: string;
   arp_no: string;
   location: string;
   assessed_value: string;
   payment_reason: string;
   total_amount: string;
   cto_no: string;
   dated: string;
   name_of_requestor: string;
   s1: string;
   s2: string;
   s3: string;
   s4: string;
   s5: string;
   verified_by: string;
   by_name1: string;
   by_title1: string;
   certification_fee: string;
   or_no: string;
   date: string;
   amount: string;
   by_name2: string;
   by_title2: string;
   remarks: string;
}